moving auth logic to another project
