+++ src/LobAccelerator.Library.Tests/appsettings.json
@@ -1,0 +1,3 @@
+﻿{
+  "": ""
+}
